9.11.

+++ Pinta.Effects/Effects/UnfocusEffect.cs
@@ -87,7 +87,7 @@ namespace Pinta.Effects
 				int alpha = sum / area;
 				int div = area * 255;
 
-				return ColorBgra.FromBgraClamped (b / div, g / div, r / div, alpha);
+				return ColorBgra.FromBgraClamped ((float)b / div, g / div, r / div, alpha);
 			}
 		}
 		#endregion
